Added Google image search.
There is a java binding for the google ajax API's
available, but the code was only obtainable from svn,
so I didn't trust that and wrote my own. It's quite
clean anyway.

I also added a test testing this functionality. It
all seems to work.

The json library was added for this purpose.