added NewConversationEvent, this event has to be sent when startConversation is called.
Also ConversationManager should listen for this event so it can handle that.
also, Account should better subclass Contact, else sending this event will be
a pain in the ass. (unless somebody comes up with a better way). the alternative
i can think of is sending our ip instead of a Contact object.