ConversationEvent does not need a Conversation to know where it came from
the key in this proces is getContact()
this will alway return "the other one" So when broadcasted localy it will return the contact you are chatting with,
the conversation can then be found through ConversationManager. When called after network passing it will return te contact where it came from.

added dummy methods to ChatApplication and Account to be able to test
only tested the local side. how to test across network?