Please make sure the tests *at least compile*.
The ./run-class.sh script cannot execute a class for some
reason (probably general ant failure) when there are tests
that do not compile. So please make sure the tests compile,
or I will come to your house in the middle of the night,
castrate you, kill your parents, and rape your sisters.